+++ NEWS.md
@@ -18,6 +18,7 @@
 * Suppress non-informative warning in `get_data()` for model data with weights.
 * `format_value()` automatically uses scientific notation for *very* large numbers (> 1e+5). Furthermore, the check for integer values was made more robust, to avoid warnings when checking *very* large numbers for integer type.
 * Improved `find_parameters()`, `get_parameters()` and `clean_parameters()` for `BFBayesFactor`-objects.
+* `get_priors()` now works for `stanmvreg` objects.
 
 ## Bug fixes